Extract logic to a module and unit test it
After playing around with a basic version of the calculator,
doing everything in the DOM handler is proving to not be optimal.

The logic of a calculator is more complex that it appears,
so let's extract all that logic to a module, split it in pure functions,
and unit test them with the nodejs built it test runner.