appendDecimalSeparator function
We're going to extract the logic for adding a decimal separator
to its own function.

We could have added it to the "appendNumber" function,
I don't think one or the other is necessary better.

We're also not extracting the logic for selecting the right operand.
This could fit into its own neatly tested function as the logic
is repeated between two functions,
however we're just not going to bother with this at this point.
If we see this code pattern repeating more then we will extract it.