Amend calculate and add tests
We can now polish our calculate function and fix it problems.

First we start by regretting our choice of having the operands stored
as string and picking "," as a decimal separator.

So we can fix the bugs around this, which creates a new problem
with handling the "Infinity" value because it gets formatted as ∞,
which can't get parsed with `parseFloat`.

We don't rewrite everything to handle all locales in this commit,
instead we double down on forcing a german locale to our calculator and we
"manually" handle the Infinity case.