Fix bugs in number formatting
These bugs are causing cascading failure,
as a badly formatted number will result in a bad first
operand for the next operation.

Mainly we handle the fact that we don't want to display "." as a
thousands separator, and we want to use scientific number notation
for numbers that are exceeding the size of our display.