Default firstOperand to 0 when setting operator
The user is seeing 0 on their display, even tho the state
of the firstOperand is undefined in the calculator.
Let's make the calculator behave like the user is expecting.

This also fixes a bug when it was not possible to set a negative number
by typing "- 4" and "Enter" for an example.
This now works as expected.